+++ lib/pack_material/entities/mr_delivery.rb
@@ -11,9 +11,11 @@ module PackMaterialApp
     attribute :driver_name, Types::String
     attribute :client_delivery_ref_number, Types::String
     attribute :delivery_number, Types::Integer
+    attribute :waybill_number, Types::Integer
     attribute :vehicle_registration, Types::String
     attribute :supplier_invoice_ref_number, Types::String
     attribute :verified, Types::Bool
+    attribute :accepted_over_supply, Types::Bool
     attribute :putaway_completed, Types::Bool
     attribute :status, Types::String
     attribute :supplier_invoice_date, Types::DateTime