+++ lib/pack_material/repositories/replenish_repo.rb
@@ -135,18 +135,23 @@ module PackMaterialApp
 
     # @return [Array] ['Purchase Order Number - from: Supplier Party Name', :id]
     def for_select_purchase_orders_with_supplier(purchase_order_id: nil)
+      purchase_orders_ds(purchase_order_id)
+        .order(:purchase_order_number)
+        .select(
+          :id,
+          :purchase_order_number,
+          Sequel.function(:fn_party_role_name, :supplier_party_role_id)
+        )
+        .map { |r| [[r[:purchase_order_number], r[:fn_party_role_name]].join(' - from: '), r[:id]] }
+    end
+
+    def purchase_orders_ds(purchase_order_id)
       if purchase_order_id
         supplier_id = DB[:mr_purchase_orders].where(id: purchase_order_id).get(:supplier_party_role_id)
-        purchase_orders = DB[:mr_purchase_orders].where(approved: true, supplier_party_role_id: supplier_id, deliveries_received: false)
+        DB[:mr_purchase_orders].where(approved: true, supplier_party_role_id: supplier_id, deliveries_received: false)
       else
-        purchase_orders = DB[:mr_purchase_orders].where(approved: true, deliveries_received: false)
+        DB[:mr_purchase_orders].where(approved: true, deliveries_received: false)
       end
-      purchase_orders = purchase_orders.order(:purchase_order_number)
-      purchase_orders.select(
-        :id,
-        :purchase_order_number,
-        Sequel.function(:fn_party_role_name, :supplier_party_role_id)
-      ).map { |r| [[r[:purchase_order_number], r[:fn_party_role_name]].join(' - from: '), r[:id]] }
     end
 
     # @return [Array] returns for select with association label name
@@ -229,6 +234,12 @@ module PackMaterialApp
       update(:mr_deliveries, id, verified: true)
     end
 
+    def accept_mr_delivery_over_supply(id)
+      waybill_no = DB[:mr_deliveries].where(id: id).get(:waybill_number)
+      update(:mr_deliveries, id, accepted_over_supply: true)
+      update_with_document_number('doc_seqs_waybill_number', id) unless waybill_no
+    end
+
     def delivery_complete_invoice(id, attrs)
       update(:mr_deliveries, id,
              invoice_error: false,
@@ -249,6 +260,10 @@ module PackMaterialApp
         return true unless has_batch
       end
       false
+    end
+
+    def items_with_over_supply(mr_delivery_id)
+      DB[:mr_delivery_items].where(mr_delivery_id: mr_delivery_id).where { quantity_over_supplied > 0 }.get(:id) # rubocop:disable Style/NumericPredicate
     end
 
     def item_has_fixed_batch(delivery_item_id)
@@ -598,12 +613,17 @@ module PackMaterialApp
     end
 
     def update_current_prices(delivery_id)
-      items = DB[:mr_delivery_items].where(mr_delivery_id: delivery_id).map { |r| { pv_id: r[:mr_product_variant_id], price: r[:invoiced_unit_price] } }
+      items = DB[:mr_delivery_items].where(mr_delivery_id: delivery_id)
+                                    .map { |r| { pv_id: r[:mr_product_variant_id], price: r[:invoiced_unit_price] } }
       items.each do |item|
-        product = config_repo.find_matres_product_variant(item[:pv_id])
-        stock_adj_price = product.stock_adj_price.positive? ? product.stock_adj_price : item[:price]
-        update(:material_resource_product_variants, item[:pv_id], current_price: item[:price], stock_adj_price: stock_adj_price)
+        update_mr_product_variant_current_prices(item)
       end
+    end
+
+    def update_mr_product_variant_current_prices(item)
+      product = config_repo.find_matres_product_variant(item[:pv_id])
+      stock_adj_price = product.stock_adj_price.positive? ? product.stock_adj_price : item[:price]
+      update(:material_resource_product_variants, item[:pv_id], current_price: item[:price], stock_adj_price: stock_adj_price)
     end
 
     def config_repo
@@ -635,16 +655,26 @@ module PackMaterialApp
     end
 
     def create_mr_delivery_item(attrs)
+      create(:mr_delivery_items, prepare_delivery_item_quantities(attrs))
+    end
+
+    # @param [Hash] attrs => quantity_received, quantity_on_note, mr_purchase_order_item_id
+    def prepare_delivery_item_quantities(attrs)
       new_attrs = add_over_under_supply_values(attrs.to_h)
-      create(:mr_delivery_items, new_attrs)
+      calculate_for_qty_returned(new_attrs)
     end
 
     # @param [Hash] attrs
     # Should only be updated if the delivery has not yet been verified
     def add_over_under_supply_values(attrs)
-      quantities                      = over_under_supply(attrs[:quantity_received], attrs[:mr_purchase_order_item_id])
-      attrs[:quantity_over_supplied]  = quantities[:quantity_over_supplied]
+      quantities = over_under_supply(attrs[:quantity_received], attrs[:mr_purchase_order_item_id])
+      attrs[:quantity_over_supplied] = quantities[:quantity_over_supplied]
       attrs[:quantity_under_supplied] = quantities[:quantity_under_supplied]
+      attrs
+    end
+
+    def calculate_for_qty_returned(attrs)
+      attrs[:quantity_returned] = attrs[:quantity_on_note] - attrs[:quantity_received]
       attrs
     end
   end