+++ lib/pack_material/test/interactors/test_mr_bulk_stock_adjustment_item_interactor.rb
@@ -5,8 +5,8 @@ require File.join(File.expand_path('../../../../test', __dir__), 'test_helper')
 module PackMaterialApp
   class TestMrBulkStockAdjustmentItemInteractor < Minitest::Test
     def test_repo
-      repo = interactor.repo
-      # repo = interactor.send(:repo)
+      # repo = interactor.repo
+      repo = interactor.send(:repo)
       assert repo.is_a?(PackMaterialApp::TransactionsRepo)
     end