+++ lib/pack_material/ui_rules/mr_delivery_rule.rb
@@ -10,9 +10,11 @@ module UiRules
       apply_form_values
 
       rules[:can_verify] = can_verify unless @mode == :new
+      rules[:can_accept] = can_accept_over_supply unless @mode == :new
       rules[:can_add_invoice] = can_add_invoice unless @mode == :new
       rules[:can_complete_invoice] = can_complete_invoice unless @mode == :new
       rules[:is_verified] = @form_object.verified if @mode == :edit
+      rules[:over_supply_accepted] = @form_object.accepted_over_supply if @mode == :edit
       rules[:invoice_completed] = @form_object.invoice_completed unless @mode == :new
       rules[:del_sub_totals] = @repo.del_sub_totals(@options[:id]) if @mode == :edit
       common_values_for_fields case @mode
@@ -52,7 +54,8 @@ module UiRules
           renderer: :select,
           prompt: true,
           options: @repo.for_select_purchase_orders_with_supplier,
-          caption: 'Purchase Order'
+          caption: 'Purchase Order',
+          sort_items: false
         }
       )
     end
@@ -116,6 +119,11 @@ module UiRules
                      else
                        delivery
                      end
+    end
+
+    def can_accept_over_supply
+      res = PackMaterialApp::TaskPermissionCheck::MrDelivery.call(:accept_over_supply, @options[:id])
+      res.success
     end
 
     def can_verify