+++ lib/pack_material/views/mr_delivery_item/show.rb
@@ -4,7 +4,7 @@ module PackMaterial
   module Replenish
     module MrDeliveryItem
       class Show
-        def self.call(id)
+        def self.call(id) # rubocop:disable Metrics/AbcSize
           ui_rule = UiRules::Compiler.new(:mr_delivery_item, :show, id: id)
           rules   = ui_rule.compile
 
@@ -15,6 +15,7 @@ module PackMaterial
               form.add_field :mr_product_variant_code
               form.add_field :quantity_on_note
               form.add_field :quantity_received
+              form.add_field :quantity_returned
               form.add_field :quantity_over_supplied
               form.add_field :quantity_under_supplied
               form.add_field :invoiced_unit_price