+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 ### Changed
 ### Fixed
 
+## [0.6.3] - 2019-07-11
+### Added
+- Deliveries can record qty returned.
+- Oversupplied deliveries must be accepted before verify can be done.
+### Changed
+- Dataminer grids now share standard javascript grid logic.
+
 ## [0.6.2] - 2019-06-27
 ### Added
 - Exceptions are emailed. The recipients and subject prefix can be configured.